requirements.txt
requests>=2.28.0
lxml>=4.9.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
